part4 d

+++ part4/blog-list-backend/controllers/blogs.js
@@ -1,6 +1,5 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
@@ -20,17 +19,12 @@ blogsRouter.get('/:id', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
-  console.log(body)
+  const user = request.user
 
   if (!body.title) {
     return response.status(400).json({ error: 'title is missing' })
   } else if (!body.url) {
     return response.status(400).json({ error: 'url is missing' })
-  }
-
-  const user = await User.findById(body.userId)
-  if (!user) {
-    return response.status(400).json({ error: 'userId missing or not valid' })
   }
 
   const blog = new Blog({
@@ -46,21 +40,34 @@ blogsRouter.post('/', async (request, response) => {
   await user.save()
 
   response.status(201).json(savedBlog)
-
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
   const id = request.params.id
+  const user = request.user
+
+  const hasUserCreated = user.blogs.includes(id)
+
+  if (!hasUserCreated) {
+    return response.status(401).json({ error: 'user is not allowed to delete this blog' })
+  }
 
   await Blog.findByIdAndDelete(id)
   response.status(204).end()
-
 })
 
 blogsRouter.put('/:id', async (request, response) => {
   const { likes } = request.body
+  const id = request.params.id
+  const user = request.user
 
-  const blog = await Blog.findById(request.params.id)
+  const hasUserCreated = user.blogs.includes(id)
+
+  if (!hasUserCreated) {
+    return response.status(401).json({ error: 'user is not allowed to delete this blog' })
+  }
+
+  const blog = await Blog.findById(id)
 
   if (!blog) {
     response.status(404).end()
